Move GpuBatchScanExec to shims
