--- sql-plugin/src/main/scala/com/nvidia/spark/GpuDataSourceV2ScanExecBase.scala
@@ -1,88 +1,0 @@
-/*
- * Copyright (c) 2021, NVIDIA CORPORATION.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.nvidia.spark
-
-import com.nvidia.spark.rapids.{GpuLeafExecNode, GpuMetric}
-
-import org.apache.spark.rdd.RDD
-import org.apache.spark.sql.catalyst.InternalRow
-import org.apache.spark.sql.catalyst.expressions.AttributeMap
-import org.apache.spark.sql.catalyst.plans.physical
-import org.apache.spark.sql.catalyst.plans.physical.SinglePartition
-import org.apache.spark.sql.catalyst.util.truncatedString
-import org.apache.spark.sql.connector.read.{InputPartition, PartitionReaderFactory, Scan, SupportsReportPartitioning}
-import org.apache.spark.sql.execution.datasources.v2.DataSourcePartitioning
-import org.apache.spark.sql.rapids.execution.TrampolineUtil
-import org.apache.spark.sql.vectorized.ColumnarBatch
-
-/**
- * GPU version of Spark's DataSourceV2ScanExecBase.
- * This is essentially verbatim but avoids the use of LeafExecNode which
- * changed hierarchy in Spark 3.2
- */
-abstract class GpuDataSourceV2ScanExecBase extends GpuLeafExecNode {
-  def scan: Scan
-
-  def partitions: Seq[InputPartition]
-
-  def readerFactory: PartitionReaderFactory
-
-  override def simpleString(maxFields: Int): String = {
-    val result =
-      s"$nodeName${truncatedString(output, "[", ", ", "]", maxFields)} ${scan.description()}"
-    TrampolineUtil.redact(sqlContext, result)
-  }
-
-  override def outputPartitioning: physical.Partitioning = scan match {
-    case _ if partitions.length == 1 =>
-      SinglePartition
-
-    case s: SupportsReportPartitioning =>
-      new DataSourcePartitioning(
-        s.outputPartitioning(), AttributeMap(output.map(a => a -> a.name)))
-
-    case _ => super.outputPartitioning
-  }
-
-  override def supportsColumnar: Boolean = {
-    require(partitions.forall(readerFactory.supportColumnarReads) ||
-        !partitions.exists(readerFactory.supportColumnarReads),
-      "Cannot mix row-based and columnar input partitions.")
-
-    partitions.exists(readerFactory.supportColumnarReads)
-  }
-
-  def inputRDD: RDD[InternalRow]
-
-  def inputRDDs(): Seq[RDD[InternalRow]] = Seq(inputRDD)
-
-  override def doExecute(): RDD[InternalRow] = {
-    val numOutputRows = longMetric(GpuMetric.NUM_OUTPUT_ROWS)
-    inputRDD.map { r =>
-      numOutputRows += 1
-      r
-    }
-  }
-
-  override def doExecuteColumnar(): RDD[ColumnarBatch] = {
-    val numOutputRows = longMetric(GpuMetric.NUM_OUTPUT_ROWS)
-    inputRDD.asInstanceOf[RDD[ColumnarBatch]].map { b =>
-      numOutputRows += b.numRows()
-      b
-    }
-  }
-}
